rename funcomics to utils
