--- liana/funcomics/__init__.py
@@ -1,4 +1,0 @@
-from liana.funcomics.obsm_to_adata import obsm_to_adata
-from liana.funcomics.mdata_to_anndata import mdata_to_anndata
-from liana.funcomics.transform import zi_minmax, neg_to_zero
-from liana.funcomics.causalnet import find_causalnet, build_prior_network
